// src/views/Voting/components/CastVoteModal/types.ts
import { InjectedModalProps } from '@pancakeswap/uikit'

export enum ConfirmVoteView {
  MAIN = 'main',
  DETAILS = 'details',
}

export interface CastVoteModalProps extends InjectedModalProps {
  onSuccess: () => Promise<void>
  proposalId: string
  vote: {
    label: string
    value: number
  }
  block?: number
  onDismiss: any
}